refactor: rename ls to list
